fix: 修复部分问题

+++ src/main/java/com/jingwei/rsswithai/application/dto/CreateModelConfigRequest.java
@@ -14,5 +14,5 @@ public class CreateModelConfigRequest {
     private Double topP;
     private Integer topK;
     private Integer maxTokens;
-    private Long seed;
+    private Integer seed;
 }
