+++ src/main/java/com/jingwei/rsswithai/application/dto/UpdateModelConfigRequest.java
@@ -11,5 +11,5 @@ public class UpdateModelConfigRequest {
     private Double topP;
     private Integer topK;
     private Integer maxTokens;
-    private Long seed;
+    private Integer seed;
 }
